requirements.txt
appnope==0.1.0
asgiref==3.2.3
backcall==0.1.0
decorator==4.4.1
Django==3.0.3
djangorestframework==3.11.0
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
Markdown==3.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
ptyprocess==0.6.0
Pygments==2.5.2
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.8
